src/examples/export/test/requirements.txt
pytest==8.4.2
cocotb==2.0.1
pillow==11.3.0
